+++ Kernel/GenericInterface/Invoker/Ticket/Generic.pm
@@ -85,7 +85,7 @@ sub PrepareRequest {
     if ( $Param{Data}->{TicketID} ) {
         %Ticket = $TicketObject->TicketDeepGet(
             TicketID                 => $Param{Data}->{TicketID},
-            ArticleID                => $Param{Data}->{ArticleID},
+            ArticleID                => $Param{Data}->{ArticleID},    # optional, hence not checked
             GetAllArticleAttachments => $GetAllArticleAttachments,
             UserID                   => 1,
         );
@@ -119,6 +119,12 @@ sub PrepareRequest {
             Data     => \%Ticket,
             HashKeys => \@HashKeys,
         );
+
+        # Also remove elements from given payload.
+        $UtilObject->DataStructureRemoveElements(
+            Data     => $Param{Data},
+            HashKeys => \@HashKeys,
+        );
     }
 
     # Base-64 encode configured field values.
@@ -135,6 +141,12 @@ sub PrepareRequest {
 
         $UtilObject->Base64DeepEncode(
             Data     => \%Ticket,
+            HashKeys => \@HashKeys,
+        );
+
+        # Also encode elements of given payload.
+        $UtilObject->Base64DeepEncode(
+            Data     => $Param{Data},
             HashKeys => \@HashKeys,
         );
     }